+++ contracts/interfaces/IMsgport.sol
@@ -19,25 +19,25 @@ pragma solidity >=0.8.9;
 
 interface IMsgport {
     event DappError(
-        uint256 _fromChainId,
+        uint64 _fromChainId,
         address _fromDappAddress,
         address _toDappAddress,
         bytes _message,
         string _reason
     );
 
-    function getLocalChainId() external view returns (uint256);
+    function getLocalChainId() external view returns (uint64);
 
     function send(
         address _throughLocalDockAddress,
-        uint256 _toChainId,
+        uint64 _toChainId,
         address _toDappAddress,
         bytes memory _messagePayload,
         bytes memory _params
     ) external payable returns (uint256);
 
     function recv(
-        uint256 _fromChainId,
+        uint64 _fromChainId,
         address _fromDappAddress,
         address _toDappAddress,
         bytes memory _message
